+++ packages/shader-composer-core/tsconfig.json
@@ -1,8 +1,4 @@
 {
   "extends": "../../tsconfig.json",
-  "include": [
-    "src/**/*",
-    "../shader-composer/src/fun.ts",
-    "../shader-composer/src/three.ts"
-  ]
+  "include": ["src/**/*"]
 }
